requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.1
bcrypt==4.3.0
beanie==1.29.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
jose==1.0.0
lazy-model==0.2.0
motor==3.7.0
passlib==1.7.4
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11.3
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
